moved capetools in with other utils
